requirements.txt
streamlit==1.24.0
python-dotenv==1.0.1
langchain==0.3.3
langchain-groq
langchain-community
langchain-google-genai
faiss-cpu